+++ HoloRepositoryUI/src/__tests__/components/MainContainer.routing.test.tsx
@@ -1,17 +1,17 @@
 import React from "react";
 import { mount, ReactWrapper } from "enzyme";
-import MainContainer from "../../components/core/MainContainer";
-import AuthContainer from "../../components/authenticated/core/AuthContainer";
+import MainContainer from "../../components/shared/MainContainer";
+import AppContainer from "../../components/app/core/AppContainer";
 import PublicContainer from "../../components/public/PublicContainer";
 import { LocationProvider, createHistory, createMemorySource } from "@reach/router";
 import LandingPage from "../../components/public/LandingPage";
-import PatientListPage from "../../components/authenticated/patients/PatientListPage";
+import PatientListPage from "../../components/app/patients/PatientListPage";
 import ErrorPage from "../../components/public/ErrorPage";
 
 it("should render PublicContainer for route '/'", () => {
   const underTest = prepareAndMountComponent("/");
   expect(underTest.find(PublicContainer)).toHaveLength(1);
-  expect(underTest.find(AuthContainer)).toHaveLength(0);
+  expect(underTest.find(AppContainer)).toHaveLength(0);
 });
 
 it("should render LandingPage for route '/start'", () => {
@@ -26,21 +26,21 @@ it("should render LandingPage for invalid URLs outside the '/app' main route", (
   expect(underTest.find(LandingPage)).toHaveLength(1);
 });
 
-it("should render AuthContainer for route '/app'", () => {
+it("should render AppContainer for route '/app'", () => {
   const underTest = prepareAndMountComponent("/app");
-  expect(underTest.find(AuthContainer)).toHaveLength(1);
+  expect(underTest.find(AppContainer)).toHaveLength(1);
   expect(underTest.find(PublicContainer)).toHaveLength(0);
 });
 
 it("should render PatientListPage for route '/app/patients'", () => {
   const underTest = prepareAndMountComponent("/app/patients");
-  expect(underTest.find(AuthContainer)).toHaveLength(1);
+  expect(underTest.find(AppContainer)).toHaveLength(1);
   expect(underTest.find(PatientListPage)).toHaveLength(1);
 });
 
 it("should render ErrorPage for invalid URLs under the '/app' main route", () => {
   const underTest = prepareAndMountComponent("/app/foobar");
-  expect(underTest.find(AuthContainer)).toHaveLength(1);
+  expect(underTest.find(AppContainer)).toHaveLength(1);
   expect(underTest.find(ErrorPage)).toHaveLength(1);
 });
