+++ HoloRepositoryUI/src/components/app/PipelineDetailBox.tsx
