+++ HoloRepositoryUI/src/components/app/core/ContentContainer.scss
