+++ HoloRepositoryUI/src/components/app/core/MenuHeader.scss
