+++ HoloRepositoryUI/src/components/app/holograms/HologramsDetailsList.tsx
