+++ HoloRepositoryUI/src/components/app/holograms/HologramsDetailsListColumns.tsx
