+++ HoloRepositoryUI/src/components/app/patient/PatientDetailPage.tsx
