+++ HoloRepositoryUI/src/components/app/patient/holograms/HologramDetailBox.tsx
