+++ HoloRepositoryUI/src/components/app/patient/holograms/HologramDetailsPanel.tsx
