+++ HoloRepositoryUI/src/components/app/patient/holograms/HologramPreview.tsx
