+++ HoloRepositoryUI/src/components/app/patient/holograms/new_hologram/pipeline/ImagingStudyDetailBox.tsx
