+++ HoloRepositoryUI/src/components/app/patient/holograms/new_hologram/upload/HealthDataFormula.tsx
