+++ HoloRepositoryUI/src/components/app/patients/PatientCard.scss
