+++ HoloRepositoryUI/src/components/app/patients/PatientCard.tsx
