+++ HoloRepositoryUI/src/components/shared/App.scss
