+++ HoloRepositoryUI/src/components/shared/MainContainer.tsx
@@ -1,15 +1,15 @@
 import React, { Component } from "react";
 import { Router } from "@reach/router";
-import AuthContainer from "../authenticated/core/AuthContainer";
+import AppContainer from "../app/core/AppContainer";
 import PublicContainer from "../public/PublicContainer";
 import AboutPage from "../public/AboutPage";
 import LandingPage from "../public/LandingPage";
 import ErrorPage from "../public/ErrorPage";
-import DeviceConnectorPage from "../authenticated/DeviceConnectorPage";
-import HologramsListPage from "../authenticated/holograms/HologramsListPage";
-import PatientListPage from "../authenticated/patients/PatientListPage";
-import PatientDetailPage from "../authenticated/patient/PatientDetailPage";
-import ProfileInformationPage from "../authenticated/ProfileInformationPage";
+import DeviceConnectorPage from "../app/DeviceConnectorPage";
+import HologramsListPage from "../app/holograms/HologramsListPage";
+import PatientListPage from "../app/patients/PatientListPage";
+import PatientDetailPage from "../app/patient/PatientDetailPage";
+import ProfileInformationPage from "../app/ProfileInformationPage";
 
 class MainContainer extends Component {
   render() {
@@ -21,14 +21,14 @@ class MainContainer extends Component {
             <AboutPage path="about" />
           </PublicContainer>
 
-          <AuthContainer path="app">
+          <AppContainer path="app">
             <PatientListPage path="patients" />
             <PatientDetailPage path="patient/:id" />
             <DeviceConnectorPage path="devices" />
             <HologramsListPage path="holograms" />
             <ProfileInformationPage path="profile" />
             <ErrorPage default />
-          </AuthContainer>
+          </AppContainer>
         </Router>
       </>
     );
