+++ HoloRepositoryUI/src/components/shared/MainFooter.tsx
